requirements.txt
ultralytics>=8.3.0
opencv-python>=4.10.0
numpy>=1.26.0
PyYAML>=6.0.1
tqdm>=4.66.0
requests>=2.31.0
